Enable draft_release, attach binaries to release, and fix static build
The fixes for the static build are working for me locally, but still
need testing in GitHub's environment.  The artifacts part has also not
been tested yet.

Not yet publishing the releases.  That step will be last, so that at
least we don't have a bunch of garbage releases on the repo.  Drafts and
tags can be more easily deleted.