Fix failing ffmpeg build script
The last command, even if it doesn't run exit 1, returns a non-zero
return code, causing the build script to fail.  This adds "true" to the
end to fix it.

As with everything else in this repo, it's untested until I push to
GitHub and try it in the Actions environment.