Add a temporary debug step
I want to know if I can refer to a local custom action from the workflow
folder without checking out the repo explicitly.  To this end, I am
checking the current working directory, github workspace directory
(should be the same), the path to the currently-running script, and
listing the contents of the directory two levels up from the current
one.

The results of this will tell me what I can get away with, and to make
it a quick test, I've added an explicit "exit 1" to make this step fail.