Custom python action to replace third-party actions
Not sure if this will work yet, as always.  But this replaces all
third-party actions with a custom action that cannot be poisoned with a
supply-chain attack.

It also adds additional features to the workflow, such as collecting the
assets attached to the release and using their MD5 sums as the body of
the release.