Additional debugging of step outputs
I have seen the release ID printed from within the custom action, and
the set-output command seems to work, insofar as it is swallowed by the
runner and does not show up in the logs.  So why is it not set when I
need it in another job?  Or in the same job?