First attempt at a Windows build
Set an environment variable for sudo ($SUDO), which will be empty on
Windows.  Not sure what else I will need to change, but hopefully I can
build all of these things in bash instead of having to resort to
separate build instructions for Windows.

Also installs nasm & yasm via chocolatey.  There may be other things
missing, but we'll start with that.