Try again on Windows without msys2/setup-msys2
My last debug session revealed that msys2/setup-msys2 installs to
D:/a/_temp/msys/msys64, and skimming the source code of that action
revealed that this is a separate install extracted from the default one
in C:/msys64.  So there is no reason it should fail if we try to use
that.  At the very least, this could bring the build environment in line
with the tmate debugging environment, which would make debugging easier
and more useful.

Ultimately, I suspect the symlinks I am creating on Windows are not
working.  But I would like to do another SSH debug session to confirm
and to develop another solution.