Rewrite the API client
The custom action API client written in Python was not able to communicate with
the GitHub servers from a Windows host due to CA cert errors.  These went away
when debugging via SSH.

This completely rewrites the API client in node using octokit, which seems to
work in all cases.