Debug GitHub Actions environment variables
The scheme I introduced in the previous commit where I collect the
various software versions into a text file and upload it as an asset
failed.  This failed because the text file uploaded from one runner
would not be able to overwrite a file of the same name.

In general, this should be a good thing.  We probably do not want an
accidental duplicate to replace something from another runner.  It would
be better for the workflow to fail.

But this means we should have another way to collect the information.  I
would prefer something that would never get out of sync with the
environment variables, so rather than reference each one in the final
job, I will see if there is a way I can collect their values in a
script.  This will fail quickly after printing the info I am interested
in.