Refactor versioning
The previous commit resulted in a successful, complete workflow... that
also leaked a token associated with shaka-bot.  Oops!  (The token has
now been burned.)

The issue was in using "env" to collect software versions.  So rather
than use the environment variables to store this information, it is now
located in a text file that makes it easy to build into our release
notes later.  When checking out code for dependencies, we use this same
file (through a shell script) to extract the version number we want to
check out.

This also changes the order of publication.  The previous workflow run
generated an email that included no real information, because we
published the release before adding any release notes.  However, the
only real dependency between publication and release notes is that we
need to publish the release before we can get MD5 sums for the binary.
So now we compile the release notes without MD5 sums, then publish, then
add the MD5 sums after publication.

Finally, this changes the token we use to the automatic GITHUB_TOKEN.  I
am not sure what the result will be, though.  What user name will be
associated with the release?  Mine?  Some GitHub bot?  Will it have all
the appropriate permissions?  My hope is that this mitigates any
possibility of a leaked, persistent token in the future.