Fix extra tags created by updating release notes
If you update the release body without specifying tag_name, it gets
reset, resulting in a new tag being created with an auto-generated name
like "untagged-SHA1".  So we need to fetch the existing name before we
update the body, and we need to specify it in the update call, too.
This is not mentioned in GitHub's docs, and may be a bug on their end.