fix: Update x264 and x265, use specific revisions
We can get x265 from git now, instead of the aptly-named mercurial.

x264 still has no tags, branches, or release numbers, but x265 does.
This now uses a specific tag of x265, and uses a specific revision of
x264 (the most recent SHA1 as of today).

This should make the build more stable.