feat: Disable nvenc on Linux
This seems to require dynamic libraries, and the static build with
nvenc does not seem to be portable.  Linux users who want hardware
acceleration with nvenc may have to use another FFmpeg binary, such as
one built from source or the one provided by their Linux distro.  With
the addition of SVT-AV1 for fast software encoding, this should not be a
huge deal.