ci: Fix ENABLE_DEBUG flag always interpreted as true
When we process the flag through JS, then print it, it becomes "true" or "false".  But we compared it to "".

The processing turns out to be necessary, since you can't refer to secrets directly when build.yaml is called from another workflow like test.yaml.  So we just fix the condition to check for "true".