ci: Use repository variables instead of secrets
This new feature of GitHub was introduced in January 2023, and gives
us an easy way to configure the repo without abusing the secrets
system.  Secrets are not available to PR contexts (for good reason),
so our secret-based configs also did not work on PRs.  By moving to
the repo variable system, we solve that problem.

https://github.blog/changelog/2023-01-10-github-actions-support-for-configuration-variables-in-workflows/