ci: Fix concurrent test runs (#49)
The concurrency group expression for the test workflow would work for
pull_request triggers, but not pull_request_target triggers. This is
because github.ref for pull_request is something like
refs/pull/<pr_number>/merge, whereas for pull_request_target triggers it
only references the target branch like refs/heads/main.

Instead we use github.event.number, which is the PR number, and fall
back to github.run_id, which should be unique for manually-triggered
runs of the workflow.