feat: Separate Linux static from hardware encoding (#48)
Hardware acceleration for Linux is too wrapped up in dynamic libraries and drivers, so from now on, builds for Linux that support hardware encoding will be dynamic and will only run on Ubuntu.  For use in every other Linux distribution, fully static builds will be done with musl in Alpine Linux.

This also updates to the latest release, FFmpeg n7.1.